feat(ccl)!: change to new incompatible schema
the changes for reissuing expiring certificates are to big to remain
backward compatibility with a reasonable amount of effort. either way,
only changes to the original reissue scenario for incorrect certificates
(e.g. 2/2 instead of 2/1) would be reflected in older releases and
changes to that logic are rather unlikely and if required could still be
downported to v2.