chore(i18n): introduce compound text keys
Texts components that are used multiple times can now be moved into
a separate text key and references in a so-called compound text key.
A compound text key must be of text type `NOTR` and this will not be
translated. In a compound text keys, other text keys can be included by
referencing them with the syntax `${OTHER_TEXT_KEY}`. The references are
replaced at build time. Text that is not a reference is preserved. This
allows for example to include line breaks, e.g. `${A}\n\n${B}`.